refactor(core): split plugin/context into setup, provides, validation
PRD #272. Replace packages/core/src/plugin/context.ts (913 LOC) with three cohesive modules:

- plugin/setup-context.ts holds createPluginSetupContext and its supporting types.
- plugin/provides-context.ts holds createPluginProvidesContext and its types/constants/helper.
- plugin/validation/* exports per-concept assert* helpers and their regex/cap constants
  (paths, identifiers, login-link, scheduled-task, meta-box-fields, component-ref), reachable
  internally by other @plumix/core modules via the validation/index.ts barrel.

The 22 register* closures stay co-located in setup-context.ts: they share the
MutablePluginRegistry argument by design, so a per-concept split would only recreate
the same coupling one layer up.

Drift-marker comments in rpc/procedures/settings/{get,schemas}.ts and
plugin/fields/repeater.ts are retargeted at the new validator paths so the
RPC-side mirror-constant follow-up still has a signpost to follow.

Public plugin-author API is unchanged: every register* signature, every error
message, every throw condition is preserved verbatim. Mechanical moves + import
updates only.

**Verification**

pnpm exec turbo run typecheck -- 29/29 packages green.
pnpm exec turbo run test lint --filter @plumix/core -- 1430/1430 tests pass, lint clean.
pnpm boundaries -- no dependency violations.

Refs #272.
