+++ packages/core/src/plugin/define.ts
@@ -1,5 +1,6 @@
-import type { PluginProvidesContext, PluginSetupContext } from "./context.js";
 import { PluginDefinitionError } from "./errors.js";
+import type { PluginProvidesContext } from "./provides-context.js";
+import type { PluginSetupContext } from "./setup-context.js";
 
 export type PluginSetup<TConfig> = (
   ctx: PluginSetupContext,
