+++ packages/core/src/plugin/index.ts
@@ -1,5 +1,6 @@
-export * from "./context.js";
 export * from "./define.js";
 export * from "./lookup.js";
 export * from "./manifest.js";
+export * from "./provides-context.js";
 export * from "./register.js";
+export * from "./setup-context.js";
