+++ packages/core/src/runtime/app.ts
@@ -6,7 +6,7 @@ import type { CapabilityResolver } from "../auth/rbac.js";
 import type { SessionPolicy } from "../auth/sessions.js";
 import type { PlumixConfig } from "../config.js";
 import type { AppContext } from "../context/app.js";
-import type { ContextExtensionEntry } from "../plugin/context.js";
+import type { ContextExtensionEntry } from "../plugin/provides-context.js";
 import type {
   PluginRegistry,
   RegisteredRawRoute,
